Bunch of improvments  - Move class SaltEventsDaemon from __init__.py to daemon.py. Fixed all usage of class to use new and correct import.  - Add new cli argument --logfile that can be used to specify logfile destination.  - Running 'salt-eventsd' from cli now do not require -l argument by default.  - Logging to console enabled when running in foreground.  - Moved __version__ string from version.py to __init__.py
