+++ salteventsd/loader.py
@@ -8,7 +8,6 @@ import sys
 import logging
 import salt.log
 import os
-import pprint
 import yaml
 
 logger = salt.log.setup.logging.getLogger(__name__)
@@ -20,7 +19,7 @@ class SaltEventsdLoader(object):
     The loader takes care of reading the configfile and
     setting up the correct logger.
     '''
-    def __init__(self, config=None):
+    def __init__(self, config=None, log_level=None, log_file=None):
         self.config_file = config if config else "/etc/salt/eventsd"
         # retrieve current settings from the config file
         self.opts = None
@@ -29,6 +28,14 @@ class SaltEventsdLoader(object):
         # make sure we have a 'general' section
         if 'general' in self.opts.keys():
             self.gen_opts = self.opts['general']
+
+        # Use log level if explicitly set from cli
+        if log_level:
+            self.gen_opts['loglevel'] = log_level
+
+        # Use log file if explicitly set from cli
+        if log_file:
+            self.gen_opts['logfile'] = log_file
 
         self._init_logger()
         log.info("loaded config from {0}".format(config))
@@ -46,16 +53,24 @@ class SaltEventsdLoader(object):
                 self.gen_opts['loglevel'],
             )
 
-        # if no log settings found, use defaults
+            salt.log.setup_console_logger(
+                log_level=self.gen_opts['loglevel'],
+            )
         else:
-            log.setup_logfile_logger('/var/log/salt/eventsd', 'warn')
+            # if no log settings found, use defaults
+            salt.log.setup_console_logger(
+                log_level="warn"
+            )
+            salt.log.setup_logfile_logger(
+                '/var/log/salt/eventsd',
+                'warn',
+            )
 
     def getopts(self):
         '''
         returns the parsed options to the SaltEventsDaemon-Class
         '''
         return self.opts
-
 
     def _read_yaml(self, path):
         '''
