--- salteventsd/version.py
@@ -1,5 +1,0 @@
-__version_info__ = (0, 9, 1)
-__version__ = '.'.join(map(str, __version_info__))
-
-if __name__ == '__main__':
-    print(__version__)
